backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
psycopg==3.1.18
psycopg-binary==3.1.18
channels==4.0.0
daphne==4.0.0
Pillow==10.2.0
django-filter==23.5
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1